% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntenyPlotteR_beta.R
\name{draw.microsynteny}
\alias{draw.microsynteny}
\title{Linear microsynteny plot}
\usage{
draw.microsynteny(
  output,
  sizefile,
  ...,
  fileformat = "png",
  colours = colours.default,
  w = 13,
  h = 5,
  opacity = 0.1,
  curve = 0.75,
  thickness = 0.5
)
}
\arguments{
\item{output}{output file name}

\item{sizefile}{Chromosome Size file}

\item{...}{synteny files (any number of alignment files can be entered)}

\item{fileformat}{output file format specified using the format \code{fileformat = "pdf"} (the default is "png")}

\item{colours}{concatenated string of chromosome IDs and assigned colours if desired using the format \code{colours = c("1" = "red", "2" = "blue", "3" = "green", "X" = "grey")} if the no colours are assigned default values will be used}

\item{w}{width of output image using the format \code{w = 13} (default)}

\item{h}{height of output image using the format \code{h = 5} (default)}

\item{opacity}{opacity of syntenic bands using the format \code{opacity = .5} (default)}

\item{curve}{degree of curve for syntenic line drawing \code{curve = .75} (default)}

\item{thickness}{thickness of the syntenic line drawing \code{thickness = .5} (default)}
}
\value{
An image file showing the linear comparison drawings
}
\description{
Draw microsynteny
}
\details{
This function draws linear synteny plots for microsyntenic regions - recommended for alignments >30K in length

It requires:

library(ggplot2) - install.packages(ggplot2)
library(ggforce) - install.packages(ggforce)
\enumerate{
\item the output file name;
\item a file with all chromosomes, chromosome lengths, and species identifiers for all species in the synteny analysis in this format:
chromosome ID, chromosome length, species identifier
\item files containing the syntenic blocks (one file per alignment, in order from most recent species alignment file to ancestor alignment file) following this format:
reference chromosome, reference start position, reference end position, target chromosome,
target start position, target end position, orient, reference species identifier, target species identifier
}

Please separate files by tab and ensure any species identifiers used between length and alignment files are matching (same identifiers and caseing)

There are optional parameters for some customization of this function:
\enumerate{
\item The format for saving the image i.e. png or pdf can be altered by inputting: fileformat = "pdf" (the default value is "png")
\item The colour of the synteny lines can be altered by inputting a concatenated string of chromosome IDs with assigned colour values which can be found with R colour Pallette
e.g. colours = c("1" = "red", "2" = "blue", "3" = "green","4" = "orange", "5" = "purple","X" = "grey") if no colours are assigned default values will be used but colours MUST be assigned to all chromosomes
\item The width of the image created can be changed by inputting: w = 5.5
\item The height of the image created can be changed by inputting: h = 10
\item the opacity of the lines can be changed by inputting: opacity = .5
\item the curvature of the lines can be changed by inputting: curve = .75
\item the line thickness can be changed by inputting: thickness = 1
}

The function works using the chromosome length file to order the Y axis and provide chromosome lengths to draw chromosome ideograms and the alignment files provides coordinates to draw the lines between ideograms

Example: draw.microsynteny("outputname","test_lengths","test_alignment_1")
}
